python-backend/requirements.txt
openai-agents
pydantic
fastapi
uvicorn
Pillow>=10.0.0
numpy>=1.24.0
